+++ Math/Mod.cs
@@ -3,11 +3,11 @@
     // Euclidian modulo: r = a - |b| floor(a/|b|)
     internal class Mod
     {
-        public static double mod(double a, double b)
+        public static decimal mod(decimal a, decimal b)
         {
             b = Absolute.abs(b);
 
-            double r = a - b * Floor.floor(a / b);
+            decimal r = a - b * Floor.floor(a / b);
             return r;
         }
     }
